+++ src/types/builder.d.ts
@@ -169,10 +169,10 @@ declare class Builder {
      * name and an attributes object to create the child it uses an
      * existing DOM element object.
      *
-     * @param {Element} elem - A DOM element.
+     * @param {Element|Builder} elem - A DOM element.
      * @return {Builder} The Strophe.Builder object.
      */
-    cnode(elem: Element): Builder;
+    cnode(elem: Element | Builder): Builder;
     /**
      * Add a child text element.
      *